Standardize types for CheckedNumerics in WebGPU https://bugs.webkit.org/show_bug.cgi?id=281294 rdar://137744198
Reviewed by Mike Wyrzykowski.

Standardize types for CheckedNumerics in WebGPU

* Source/WebGPU/WebGPU/Buffer.mm:
(WebGPU::computeRangeSize):
* Source/WebGPU/WebGPU/CommandEncoder.mm:
(WebGPU::validateResolveQuerySet):
* Source/WebGPU/WebGPU/RenderBundleEncoder.mm:
(WebGPU::RenderBundleEncoder::drawIndexed):
* Source/WebGPU/WebGPU/RenderPassEncoder.mm:
(WebGPU::RenderPassEncoder::clampIndexBufferToValidValues):
(WebGPU::RenderPassEncoder::drawIndexed):

Canonical link: https://commits.webkit.org/285017@main

+++ Source/WebGPU/WebGPU/Buffer.mm
@@ -245,7 +245,7 @@ bool Buffer::validateGetMappedRange(size_t offset, size_t rangeSize) const
 
 static size_t computeRangeSize(uint64_t size, size_t offset)
 {
-    auto result = checkedDifference<size_t>(size, offset);
+    auto result = checkedDifference<uint64_t>(size, offset);
     if (result.hasOverflowed())
         return 0;
     return result.value();
